chore: move WaitingPlayerList ui FreeMatch page의 WaitingPlayer ui segment를 Widget 의 Slice로 변경
