+++ src/2_Widget/WaitingPlayerList/index.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import Button from "../../../../4_Shared/components/Button";
-import useGetWaitingPlayerList from "../../../../3_Entity/Match/useGetWaitingPlayerList";
+import Button from "../../4_Shared/components/Button";
+import useGetWaitingPlayerList from "../../3_Entity/Match/useGetWaitingPlayerList";
 import WaitingPlayerCard from "./ui/WaitingPlayerCard";
-import useInfiniteScrollPaging from "../../../../4_Shared/model/useInfiniteScrollPaging";
-import ronaldoBanner from "../../assets/svg/ronaldoBanner.svg";
+import useInfiniteScrollPaging from "../../4_Shared/model/useInfiniteScrollPaging";
+import ronaldoBanner from "../../4_Shared/assets/svg/ronaldoBanner.svg";
 const WaitingPlayerList = () => {
   const [page, setPage] = React.useState<number>(1);
   const [isWatingPlayerListOpen, setIsWaitingPlayerListOpen] =
