+++ src/2_Widget/WaitingPlayerList/ui/WaitingPlayerCard/type.ts
